More build fixes.
#rb none
#lockdown Nick.Penwarden

[CL 3234301 by Ben Marsh in Main branch]

+++ Engine/Plugins/Online/OnlineSubsystemOculus/Source/Private/OnlineSubsystemOculusTypes.h
@@ -96,7 +96,7 @@ public:
 	/** Needed for TMap::GetTypeHash() */
 	friend uint32 GetTypeHash(const FUniqueNetIdOculus& A)
 	{
-		return GetTypeHash(A.ID);
+		return GetTypeHash((uint64)A.ID);
 	}
 };
